feat: Refactoring Tables - Adding reusable Application columns, so we can use it in the import and main grid - Fixing issue with borders within the grid - Fixing how the attributes display in ExerciseActivities - Adding the search to the pagination table - Fixing issue with getColumnSortIdentity and adding it to the table instance
